feat: YML manifest. Simplify YBP. (#2353)
This contains some fairly hefty changes to how we build.
* `build_manifest.json` is now `build_manifest.yml` (leaner and we can
add comments!)
* `build_manifest.sh` is gone. `bootstrap_docker.sh` knows what to build
depending on what you ask it to build. *WIP*
  * Fixup args.
* Remove some unneeded installs from
`barretenberg/cpp/dockerfiles/Dockerfile.wasm-linux-clang`
* `jq` removed from repo as installed by default in ubuntu. But... `yq`
isn't on our spot AMI, so adding for now.
* We shift to using docker buildkit with `DOCKER_BUILDKIT=1` being set.
Needed so we can have different `.dockerignore` files for the same build
dir. Also it's just the newer, faster way.
* `yarn-project-base` now doesn't need to have every workspace listed in
it. The dockerignore file ensures only the package.json and
tsconfig.json files are in the context.
* `yarn-project` now has a Dockerfile at root that builds upon the
above. It includes the full workspace, builds it, checks formatting, and
runs all the unit tests in parallel on a 64 core machine (e2e and canary
tests excluded).
* Various tweaks to various files to make formatting, tests etc pass in
this way.

---------

Co-authored-by: ludamad <adam.domurad@gmail.com>