chore: CLI's startup time was pushing almost 2s. This gets the basic 'help' down to 0.16. (#3529)
* The CLI took 2s just to display the help (worse when launching via
docker). Pretty frustrating.
* This is symptomatic of just a deeper issue, of exporting a lot of
stuff, through a single import module, or just having too much code in
one place.
* This improves our cli, getting simple printing of help down to ~0.16s.
When running in docker we get a flat overhead of about 0.4s, giving us
~0.6s, which feels responsive enough. (~0.8 on mac)
* This is achieved via several techniques. Given that `cli`
"does-almost-all-the-things", it basically needs to compile the entire
codebase on startup if we naively import everything.
* I've put every command action code into it's own file named after that
action, under the `cmds` folder. This is just nicer anyway as we
shouldn't be creating multiple 1000's line files.
* The action file is then dynamically imported if that action is
requested, so at that point in time, node will incur the import
overhead.
* In `utils` (please let's not name things utils, break it up into
distinct groups with well named files), we use a new "granular" export
api from `aztec.js`. The api isn't complete, it was just enough to
import what was needed and get rid of about 200ms of overhead.
* I added a `preAction` hook which is when we call `initAztecJs`, a new
method which triggers 1-200ms if initialization in `bb.js`. This is now
also called at start of sandbox. If running in jest, this is
automatically called on module load, to save having to mess with a
gazillion test files.
* Reworked the bb.js portal workflow so that changes are immediately
reflected in yarn-project. We produced a `package` before as part of
build, but meant this step needed to be run after each change. Now the
`package` (what would be published to npm), is just used as part of the
Docker build.
* Reworked a bit of how bb.js singleton is initialised and subsequently
used.
* Added a `./build-system/start_interactive` script, to drop you into a
shell with environment configured appropriately for running build-system
commands. Useful for debugging rebuild patterns etc. See screenshot.
* As mentioned above, start of a granular 'aztec.js` import api. This
provides still provides an abstraction over what our internal module
structure is like. We can structure the api however we like, it just
provides an alternative to importing everything at once.
* Fixed insufficient signal handling on the sandbox/cli. Sandbox set
signal handlers too late meaning the app would only respond to sigint
etc if it succesfully reached a certain point. Now installs handlers
early, and replaces later for graceful shutdown. (You wouldn't have seen
this via yarn as yarn handles signals itself).
* Compile our noir-contracts over all cores, not just 4.


![image](https://github.com/AztecProtocol/aztec-packages/assets/5764343/cd1a6089-aa82-43fd-a65c-d779fc401ea7)